Move payload span helpers under worker
